Update Bonus.java
Changes made:
1- put comments
2- changed double in bonus.java for warning Block.getWidth() / 2': integer division in floating-point context

+++ CourseworkGame-master/CourseworkGame-master/src/main/java/brickGame/Bonus.java
@@ -8,20 +8,29 @@ import java.io.Serializable;
 import java.util.Random;
 
 public class Bonus implements Serializable {
+    // type of object --> gives bonus
     public Rectangle choco;
 
+    // x and y coordinates of brick
     public double x;
     public double y;
+
+    // the time when brick was created
     public long timeCreated;
+
+    // boolean field to represent if bonus has been taken
     public boolean taken = false;
 
     public Bonus(int row, int column) {
-        x = (column * (Block.getWidth())) + Block.getPaddingH() + (Block.getWidth() / 2) - 15;
-        y = (row * (Block.getHeight())) + Block.getPaddingTop() + (Block.getHeight() / 2) - 15;
+
+        // draw the bonus object after setting the x and y position on grid for block
+        x = (column * (Block.getWidth())) + Block.getPaddingH() + ((double) Block.getWidth() / 2) - 15;
+        y = (row * (Block.getHeight())) + Block.getPaddingTop() + ((double) Block.getHeight() / 2) - 15;
 
         draw();
     }
 
+    // creates the bonus object
     private void draw() {
         choco = new Rectangle();
         choco.setWidth(30);
@@ -30,15 +39,15 @@ public class Bonus implements Serializable {
         choco.setY(y);
 
         String url;
-        if (new Random().nextInt(20) % 2 == 0) {
+        // based on  result of if statement, sets picture to bonus1 or bonus2
+        // introduces randomness
+
+        if (new Random().nextInt(20) % 2 == 0) { // if random number is even
             url = "bonus1.png";
-        } else {
+        } else {                    // if randomly generated number is odd
             url = "bonus2.png";
         }
 
         choco.setFill(new ImagePattern(new Image(url)));
     }
-
-
-
 }
